+++ src/getValueProviderImplementations.mjs
@@ -2,13 +2,16 @@
 
 /**
  * @param {string} env_previx
+ * @param {boolean | null} cli_param
  * @returns {Promise<ValueProviderImplementation[]>}
  */
-export async function getValueProviderImplementations(env_previx) {
+export async function getValueProviderImplementations(env_previx, cli_param = null) {
     return [
-        (await import("./ValueProviderImplementation/CliParamValueProviderImplementation.mjs")).CliParamValueProviderImplementation.new(
-            process.argv
-        ),
+        ...cli_param ?? true ? [
+            (await import("./ValueProviderImplementation/CliParamValueProviderImplementation.mjs")).CliParamValueProviderImplementation.new(
+                process.argv
+            )
+        ] : [],
         (await import("./ValueProviderImplementation/EnvValueProviderImplementation.mjs")).EnvValueProviderImplementation.new(
             process.env,
             env_previx