+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-03-20-1
+
+Changes:
+
+- Optional cli param
+
 ## v2023-03-16-1
 
 Changes: